refactor(queryflux): extract bootstrap/serve into lib.rs behind a builder API
Move construction and serve/shutdown out of main.rs into a new
QueryFlux::builder().build().await?.serve().await surface. The shipped
binary becomes a thin consumer of the same builder; behavior is
unchanged otherwise (tracing init switched from .init() to .try_init()
so embedders can construct multiple QueryFlux instances in one process,
e.g. in tests, without panicking).

Part of #198 (PR 1/4: library extract).
